+++ Controllers/UnidadeMedidaController.cs
@@ -2,6 +2,9 @@
 using Microsoft.EntityFrameworkCore;
 using PetApi.Data;
 using PetApi.Models;
+using PetApi.ViewModels;
+using PetApi.ViewModels.SubcategoriaVM;
+using PetApi.ViewModels.UnidadeMedidaVM;
 
 namespace PetApi.Controllers;
 
@@ -23,58 +26,79 @@ public class UnidadeMedidaController : ControllerBase
         try
         {
             if (_context.UnidadeMedidas == null)
-            {
-                return NotFound();
-            }
-            return await _context.UnidadeMedidas.ToListAsync();
-        }
-        catch (Exception e)
-        {
-            return StatusCode(500, e);
-        }
+                return NotFound(new ResultViewModel<UnidadeMedida>("Não existe nenhuma unidade medida"));
 
+            var unidadeMedida = await _context.UnidadeMedidas.ToListAsync();
+
+            return Ok(new ResultViewModel<List<UnidadeMedida>>(unidadeMedida));
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<UnidadeMedida>("Falha interna do servidor"));
+        }
     }
     // GET BY ID:
     [HttpGet("{id}")]
     public async Task<ActionResult<UnidadeMedida>> GetUnidadeMedida(int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<UnidadeMedida>("Falha interna no servidor"));
         try
         {
-            var UnidadeMedida = await _context.UnidadeMedidas.FindAsync(id);
+            var unidadeMedida = await _context.UnidadeMedidas.FindAsync(id);
+            if (unidadeMedida == null)
+                return NotFound(new ResultViewModel<SubCategoria>("Não existe essa unidade de medida"));
 
-            if (UnidadeMedida == null)
-                return NotFound();
-
-            return UnidadeMedida;
+            return Ok(new ResultViewModel<UnidadeMedida>(unidadeMedida));
         }
         catch
         {
-            return StatusCode(500);
+            return StatusCode(500, new ResultViewModel<UnidadeMedida>("Falha interna do servidor"));
         }
-
     }
-
     // POST
     [HttpPost]
-    public async Task<ActionResult<UnidadeMedida>> PostUnidadeMedida(UnidadeMedida unidadeMedida)
+    public async Task<ActionResult<UnidadeMedida>> PostUnidadeMedida(CreateUnidadeMedidaVM unidadeMedidaVM)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<UnidadeMedida>("Falha interna no servidor"));
+
         if (_context.UnidadeMedidas == null)
-            return Problem("Entidade adiciona 'ApplicationDbContext.Categoria' é nula.");
+            return NotFound((new ResultViewModel<UnidadeMedida>("Unidade de medida é nula")));
+        try
+        {
+            var unidadeMedida = new UnidadeMedida
+            {
+                UnidadeMedidaId = 0,
+                Ativo = 0,
+                Nome = unidadeMedidaVM.Nome,
+                Sigla = unidadeMedidaVM.Sigla
+            };
 
-        _context.UnidadeMedidas.Add(unidadeMedida);
-        await _context.SaveChangesAsync();
+            _context.UnidadeMedidas.Add(unidadeMedida);
+            await _context.SaveChangesAsync();
 
-        return CreatedAtAction("GetUnidadeMedida", new { id = unidadeMedida.UnidadeMedidaId }, unidadeMedida);
+            return CreatedAtAction("GetUnidadeMedida", new { id = unidadeMedida.UnidadeMedidaId }, unidadeMedida);
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<UnidadeMedida>("Falha interna do servidor"));
+        }
+        
     }
-
     // PUT:
     [HttpPut("{id}")]
-    public async Task<IActionResult> Put(int id, UnidadeMedida unidadeMedida)
+    public async Task<IActionResult> Put(int id, UpdateSubcategoriaVM unidadeMedidaVM)
     {
-        if (id != unidadeMedida.UnidadeMedidaId)
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<UnidadeMedida>("Falha interna no servidor"));
+
+        var unidadeMedida = new UnidadeMedida
         {
-            return BadRequest("01xE4 - Id diferente do Id da Unidade Medida");
-        }
+            UnidadeMedidaId = id,
+            Ativo = unidadeMedidaVM.Ativo,
+            Nome = unidadeMedidaVM.Nome
+        };
 
         _context.Entry(unidadeMedida).State = EntityState.Modified;
 
@@ -85,34 +109,41 @@ public class UnidadeMedidaController : ControllerBase
         catch (DbUpdateConcurrencyException)
         {
             if (!UnidadeMedidaExists(id))
-                return NotFound("");
+                return NotFound(new ResultViewModel<UnidadeMedida>("Unidade de medida não existe"));
 
-
-            return BadRequest("01xE5 - Falha ao alterar categoria");
-
+            return BadRequest(new ResultViewModel<UnidadeMedida>("Falha ao alterar a unidade de medida"));
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<UnidadeMedida>("Falha interna do servidor"));
         }
 
-        return NoContent();
+        return Ok(new ResultViewModel<UnidadeMedida>("Unidade de medida alterada"));
     }
 
     //Delete
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteUnidadeMedida(int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<UnidadeMedida>("Falha interna no servidor"));
+            
         if (_context.UnidadeMedidas == null)
-        {
-            return NotFound();
-        }
-        var unidadeMedida = await _context.UnidadeMedidas.FindAsync(id);
-        if (unidadeMedida == null)
-        {
-            return NotFound();
-        }
+            return NotFound(new ResultViewModel<UnidadeMedida>("Unidade de medida é nula"));
 
-        _context.UnidadeMedidas.Remove(unidadeMedida);
-        await _context.SaveChangesAsync();
-
-        return NoContent();
+        try
+        {
+            var unidadeMedida = await _context.UnidadeMedidas.FindAsync(id);
+            if (unidadeMedida == null)
+                return NotFound(new ResultViewModel<UnidadeMedida>("Unidade de medida é não encotrada"));
+            _context.UnidadeMedidas.Remove(unidadeMedida);
+            await _context.SaveChangesAsync();
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<UnidadeMedida>("Falha interna do servidor"));
+        }
+        return Ok(new ResultViewModel<UnidadeMedida>("Unidade de medida excluída"));
     }
     private bool UnidadeMedidaExists(int id)
     {
